Move types from state.ts to types.ts
