+++ app/frontend/components/pages/accounts/accountsDashboard.tsx
@@ -2,14 +2,14 @@ import {Fragment, h} from 'preact'
 import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import printAda from '../../../helpers/printAda'
-import {Lovelace, State} from '../../../state'
+import {State} from '../../../state'
 import {AdaIcon} from '../../common/svg'
 import Alert from '../../common/alert'
 import SendTransactionModal from './sendTransactionModal'
 import DelegationModal from './delegationModal'
 import ConfirmTransactionDialog from '../../../../frontend/components/pages/sendAda/confirmTransactionDialog'
 import AccountTile from './accountTile'
-import {AccountInfo} from '../../../../frontend/types'
+import {AccountInfo, Lovelace} from '../../../../frontend/types'
 import Conversions from '../../common/conversions'
 
 type DashboardProps = {
