+++ app/frontend/helpers/printAda.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-unused-vars
-import {Lovelace} from '../state'
+import {Lovelace} from '../types'
 export default (coins: Lovelace, decimals = 6): string => {
   const adaAmount = coins * 0.000001
